requirements.txt
openai
chromadb
playwright
beautifulsoup4
requests
tqdm
python-dotenv
